add Utils

+++ code/Math/Exgcd.cpp
@@ -4,7 +4,7 @@ i64 exgcd(i64 a, i64 b, i64 &x, i64 &y) {
         x = 1, y = 0;
         return a;
     }
-    int g = exgcd(b, a % b, y, x);
+    i64 g = exgcd(b, a % b, y, x);
     y -= a / b * x;
     return g;
 }
